Add files via upload
The Iris flower dataset is a popular and widely used dataset in machine learning and statistics. It was introduced by the British statistician and biologist Ronald Fisher in 1936. The dataset is sometimes referred to as Fisher's Iris dataset or Anderson's Iris dataset.

The dataset consists of measurements of four features of three different species of Iris flowers: Setosa, Versicolor, and Virginica. The four features are:

1. Sepal length (in centimeters)
2. Sepal width (in centimeters)
3. Petal length (in centimeters)
4. Petal width (in centimeters)

For each species of Iris, Fisher collected 50 samples, resulting in a total of 150 samples in the dataset. The goal of using this dataset is typically to classify new observations based on the measurements of these features.

The Iris dataset is commonly used for teaching and learning purposes in the fields of statistics, machine learning, and pattern recognition. It serves as a simple yet versatile dataset for tasks such as classification, clustering, and visualization.

You can find the Iris dataset in various machine learning libraries and repositories, such as the UCI Machine Learning Repository. It has become a standard benchmark dataset in the field and is frequently used to evaluate and compare the performance of different machine learning algorithms.